경험치 슬라이더 오류 수정

+++ Assets/Script/MainMenu/HUDController.cs
@@ -86,7 +86,7 @@ public class HUDController : MonoBehaviour {
     
     public void SetResourcesUI() {
         lvValue.text = AccountManager.Instance.userResource.lv.ToString();
-        expSlider.fillAmount = AccountManager.Instance.userResource.exp / AccountManager.Instance.userResource.lvExp;
+        expSlider.fillAmount = (float)AccountManager.Instance.userResource.exp / (float)AccountManager.Instance.userResource.lvExp;
         expValueText.text = AccountManager.Instance.userResource.exp.ToString() + "/" + AccountManager.Instance.userResource.lvExp;
         crystalValue.text = AccountManager.Instance.userResource.crystal.ToString();
         goldValue.text = AccountManager.Instance.userResource.gold.ToString();
